+++ test/test_KubeLibrary.py
@@ -395,14 +395,17 @@ class TestKubeLibrary(unittest.TestCase):
 
     @responses.activate
     def test_KubeLibrary_dynamic_create(self):
+        with open('test/resources/pod.json') as json_file:
+            sample_pod = json.load(json_file)
         responses.add("GET", "/version", status=200)
         responses.add("GET", "/apis", status=200, body='{"groups": [], "kind": "Pod" }', content_type="application/json")
         responses.add("GET", "/api/v1", status=200,
                       body='{"resources": [{"api_version": "v1", "kind": "Pod", "name": "Mock"}], "kind": "Pod"}',
                       content_type="application/json")
-        responses.add("POST", "/api/v1/mock", status=200)
+        responses.add("POST", "/api/v1/mock", status=200, body=json.dumps(sample_pod), content_type="application/json")
         kl = KubeLibrary(kube_config='test/resources/k3d')
-        kl.create("v1", "Pod", name="Mock")
+        created = kl.create("v1", "Pod", body=json.dumps(sample_pod))
+        self.assertEqual(created.to_dict(), sample_pod)
 
     @responses.activate
     def test_KubeLibrary_dynamic_delete(self):